fix(nodejs): Log the selected listen port
Currently, the selected port from the nodejs template gets
logged into the terminal, but in case that a `PORT` is specified
as an environment variable, the port from template is still logged
making it misleading.

This commit provides an alternative to the previous situation by
first deciding which port will be used and after that logging the
selected port.